% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPTH.R
\name{DEPTH}
\alias{DEPTH}
\title{Evaluation of the tumor heterogeneity level.}
\usage{
DEPTH(exp, match)
}
\arguments{
\item{exp}{gene expression profiles in both tumor and normal samples (microarray or RNA-Seq data, log2 transformed is recommended).
DEPTH function support 2 kind of datasets:
1) if the "exp" dataset has normal samples and tumor samples, we can caculate the heterogeneity score with both normal samples and tumor samples.
2) else if the "exp" dataset only has tumor samples, we can caculate the heterogeneity score by with tumor samples.}

\item{match}{samples in "exp" denoted by "Tumor" or "Normal".}
}
\value{
A matrix with 2 columns:
\item{Sample}{Tumor samples to be predicted.}
\item{heterogeneity score}{The heterogeneity score of each sample.}
}
\description{
\code{DEPTH} DEPTH evaluates the tumor heterogeneity level of each tumor sample based on gene expression profiles Heterogeneity score vignette.
}
\details{
DEPTH evaluates the tumor heterogeneity level of each tumor sample based on gene expression profiles.
}
\author{
Xiaosheng Wang <xiaosheng.wang@cpu.edu.cn>
}
