\name{DEPTH}
\alias{DEPTH}
\title{DEPTH evaluates the tumor heterogeneity level of each tumor sample based on gene expression profiles Heterogeneity score vignette}
\usage{
DEPTH()
}
\description{
DEPTH evaluates the tumor heterogeneity level of each tumor sample based on gene expression profiles Heterogeneity score vignette.
}
\examples{
DEPTH(exp, match)
}
